Assignment_7/samples/A7E3.py
import sys

# Module docstring
"""
This script calculates total seconds from hours, minutes, and seconds passed as command line arguments.
"""

# Constants
SEC_PER_MIN = 60

# Main Body
def main():
    main()

# Functions
def calc_total_seconds(hours, minutes, seconds):
    """ Calculate total seconds from hours, minutes, and seconds """
    try:
        total_seconds = int(hours) * 3600 + int(minutes) * SEC_PER_MIN + int(seconds)
        return total_seconds
    except:
        return None

# Main script
if __name__ == "__main__":
    main()
